% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_preproc.R
\name{basic_preproc}
\alias{basic_preproc}
\title{Transform an AnnData input to one that has had gene filtering,
normalization, log transformation, low-variance gene exclusion,
and discretization applied.}
\usage{
basic_preproc(
  ad,
  workdir = tempdir(),
  min_n_cells = 3L,
  norm_method = "lib_size",
  do_log = TRUE,
  n_top_genes = NULL,
  n_bins = 6L,
  simba_ref
)
}
\arguments{
\item{ad}{AnnData instance (i.e., inherits from `anndata._core.anndata.AnnData`)}

\item{workdir}{character(1) defines working directory for graph serialization; set to NULL to take simba default}

\item{min_n_cells}{integer(1) used to filter genes, excluding those expressed in fewer than `min_n_cells` cells}

\item{norm_method}{character(1) defaults to 'lib_size', see simba doc}

\item{do_log}{logical(1) if TRUE, conducts log normalization}

\item{n_top_genes}{NULL or numeric(1) if non-null, genes are excluded if
variance is smaller than that obtained for the "nth" top}

\item{n_bins}{integer(1) number of bins for expression discretization}

\item{simba_ref}{instance of python.builtin.module, checked to have component 'tl'}
}
\value{
instance of AnnData with layers raw and simba, the latter consisting
of a sparse matrix, and other components as determined by the operations
selected through argument bindings.
}
\description{
Transform an AnnData input to one that has had gene filtering,
normalization, log transformation, low-variance gene exclusion,
and discretization applied.
}
