% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_acts.R
\name{build_and_train_pbg}
\alias{build_and_train_pbg}
\title{build and train SIMBA PBG graph}
\usage{
build_and_train_pbg(
  adSimb,
  copy = FALSE,
  use_hvg = FALSE,
  dirname = "graph0",
  auto_wd = TRUE,
  save_wd = TRUE,
  output = "model",
  simba_ref
)
}
\arguments{
\item{adSimb}{an instance of AnnData with simba layer as
produced by basic_preproc}

\item{copy}{logical(1) passed to gen_graph in simba.tl module, defaults
to FALSE}

\item{use_hvg}{logical(1) passed as `use_highly_variable` to gen_graph, defaults to FALSE}

\item{dirname}{character(1) defaults to 'graph0'}

\item{auto_wd}{logical(1) defaults to TRUE, for automatic setting of weight decay}

\item{save_wd}{logical(1) defaults to TRUE}

\item{output}{character(1) defaults to 'model'}

\item{simba_ref}{instance of python.builtin.module, checked to have component 'tl'}
}
\description{
build and train SIMBA PBG graph
}
\examples{
p15 = get_paul15()
ref = simba_ref()
pp = ref$read_h5ad(p15)
bb = basic_preproc(pp, simba_ref=ref)
gg = build_and_train_pbg( bb, simba_ref=ref )
dict = ref$read_embedding()
dict 
}
