% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{compare_entities}
\alias{compare_entities}
\title{use the compare_entities method to assess 'closeness' of embedded elements}
\usage{
compare_entities(simbaref, refemb, queryemb, n_top = 50L, temperature = 1)
}
\arguments{
\item{simbaref}{instance of python module for simba}

\item{refemb}{AnnData instance with 'reference' embedding}

\item{queryemb}{AnnData instance with 'query' embedding}

\item{n_top}{integer(1) number of 'reference' elements to
be examined for 'max' output summarizing closeness (maximum of normed dot product)}

\item{temperature}{numeric(1), a tuning parameter whose reciprocal scales softmax
transformations of dot products between reference and query entities, defaulting
to 1.0 for this application}
}
\description{
use the compare_entities method to assess 'closeness' of embedded elements
}
\examples{
  simbaref = simba_ref()
  cemb = simbaref$read_h5ad( get_3k_cell_emb() )
  gemb = simbaref$read_h5ad( get_3k_gene_emb() )
  cc = compare_entities( simbaref, cemb, gemb )
  cc
  head(cc$var)
}
