% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_acts.R
\name{flatten_json}
\alias{flatten_json}
\title{Helper function to recursively flatten nested lists
produced by Chat-GPT 3}
\usage{
flatten_json(x, prefix = "")
}
\description{
Helper function to recursively flatten nested lists
produced by Chat-GPT 3
}
