% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paul15.R
\name{get_10x3kpbmc_path}
\alias{get_10x3kpbmc_path}
\title{retrieve path to 10x 3k PBMC h5ad (compressed)}
\usage{
get_10x3kpbmc_path(...)
}
\arguments{
\item{\dots}{passed to R.utils::decompressFile, note that ext, FUN,
temporary and remove are set.}
}
\description{
retrieve path to 10x 3k PBMC h5ad (compressed)
}
