% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paul15.R
\name{get_3k_cell_emb}
\alias{get_3k_cell_emb}
\title{retrieve path to the 3k PBMC cell embedding in AnnData format}
\usage{
get_3k_cell_emb()
}
\description{
retrieve path to the 3k PBMC cell embedding in AnnData format
}
