% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbg_ops.R
\name{get_emb_h5}
\alias{get_emb_h5}
\title{retrieve and HDF5 representation of an embedding produced by pbg in simba}
\usage{
get_emb_h5(
  ctarpath,
  type = "cell",
  numtag = "_0",
  gname = "graph0",
  mname = "model",
  spec = "v10",
  exdir = tempdir()
)
}
\arguments{
\item{ctarpath}{character(1) path to a compressed tar of pbg folder}

\item{type}{character(1) 'cell' or 'gene'}

\item{numtag}{character(1) defaults to "_0"}

\item{gname}{character(1) used as dirname in build_and_train_pbg}

\item{mname}{character(1) used as output in build_and_train_pbg}

\item{spec}{character(1) defaults to "v10"}

\item{exdir}{character(1) path where tar operation outputs will be placed,
defaults to tempdir()}
}
\value{
result of rhdf5::h5read on the embedding values
}
\description{
retrieve and HDF5 representation of an embedding produced by pbg in simba
}
\examples{
tpath = system.file(file.path("extdata", "pbg3k.tar.xz"), package="Simba4Bioc")
cemb = get_emb_h5(tpath)
dim(cemb)
}
