% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paul15.R
\name{get_paul15_path}
\alias{get_paul15_path}
\title{retrieve path to Paul 2015 (Cell 163:7, 1663) AnnData instance}
\usage{
get_paul15_path(...)
}
\arguments{
\item{\dots}{passed to R.utils::decompressFile, note that ext, FUN,
temporary and remove are set.}
}
\description{
retrieve path to Paul 2015 (Cell 163:7, 1663) AnnData instance
}
