% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{scanpy_ref}
\alias{scanpy_ref}
\title{hand scanpy from specified basilisk environment}
\usage{
scanpy_ref()
}
\description{
hand scanpy from specified basilisk environment
}
\note{
assumed implicitly installed with simba 1.2
}
