% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode.R
\name{simba_barcode_data}
\alias{simba_barcode_data}
\title{'barcode' data for figure 2d of original simba paper}
\usage{
simba_barcode_data(
  symbols,
  origad,
  clabel,
  simbaref,
  cembpath,
  gembpath,
  layer = "softmax"
)
}
\arguments{
\item{symbols}{character() vector of gene symbols for barcode visualization}

\item{origad}{AnnData reference to original RNA-seq data}

\item{clabel}{character(1) column name for entity type in origad}

\item{simbaref}{instance of python module for simba}

\item{cembpath}{character(1) path to h5ad file for cell embedding}

\item{gembpath}{character(1) path to h5ad file for gene embedding}

\item{layer}{character(1) either 'softmax' (default) or 'norm', used
as source of scoring of strength of cell-gene association}
}
\description{
'barcode' data for figure 2d of original simba paper
}
\examples{
sref = simba_ref()
p3k = sref$read_h5ad(get_10x3kpbmc_path(overwrite=TRUE))
myd = simba_barcode_data( symbols = c("MS4A1", "CST3", "NKG7"),
   origad = p3k, clabel = "celltype",
   simbaref = sref, cembpath = get_3k_cell_emb(),
   gembpath = get_3k_gene_emb() )
head(myd,3)
table(myd$symbol)
}
