% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_acts.R
\name{ingest_training_stats}
\alias{ingest_training_stats}
\title{get pbg_metrics data frames}
\usage{
ingest_training_stats(simba_ref, gg_dirname = "graph0", tr_output = "model")
}
\arguments{
\item{simba_ref}{reference to simba module}

\item{gg_dirname}{character(1) the dirname specified for build_and_train_pbg,
defaults to 'graph0'}

\item{tr_output}{character(1) the output specified for build_and_train_pbg
defaults to 'model'}
}
\value{
a list with two data.frames of training statistics
}
\description{
get pbg_metrics data frames
}
\examples{
p15 = get_paul15(overwrite=TRUE) # allow repetition
ref = simba_ref()
pp = ref$read_h5ad(p15)
bb = basic_preproc(pp, simba_ref=ref)
gg = build_and_train_pbg( bb, simba_ref=ref )
ts = ingest_training_stats(ref)
lapply(ts, head, 2)
}
