% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_acts.R
\name{propagate_label}
\alias{propagate_label}
\title{try propagation}
\usage{
propagate_label(simba_ref, origad, label_name)
}
\description{
try propagation
}
\examples{
p15 = get_paul15_path(overwrite=TRUE) # allow repetition
ref = simba_ref()
pp = ref$read_h5ad(p15)
bb = basic_preproc(pp, simba_ref=ref)
gg = build_and_train_pbg( bb, simba_ref=ref )
labc = propagate_label( ref, pp, "paul15_clusters" )
}
